Changes day from float to integer

+++ database/migrations/20190116223230_hours_of_operation.js
@@ -1,7 +1,7 @@
 exports.up = knex =>
   knex.schema.createTable('hours_of_operation', table => {
     table.uuid('id').primary()
-    table.float('day').notNullable()
+    table.integer('day').notNullable()
     table.float('open_time').notNullable()
     table.float('close_time').notNullable()
     table.boolean('closed')
